requirements.txt
# Zord Coder v1 - Core Requirements
# Essential dependencies only

llama-cpp-python>=0.2.0
rich>=13.0.0
huggingface-hub>=0.19.0
